target/s390x: Add a CONFIG switch to disable legacy CPUs
RH-Author: Thomas Huth <thuth@redhat.com>
RH-MergeRequest: 252: s390x: remove legacy CPU types
RH-Jira: RHEL-39898
RH-Acked-by: Cédric Le Goater <clg@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Commit: [3/5] f8e78c8e0349c8645e7df7b0bebed1635865b454 (thuth/qemu-kvm-cs9)

The oldest model that IBM still supports is the z13. Considering
that each generation can "emulate" the previous two generations
in hardware (via the "IBC" feature of the CPUs), this means that
everything that is older than z114/196 is not an officially supported
CPU model anymore. The Linux kernel still support the z10, so if
we also take this into account, everything older than that can
definitely be considered as a legacy CPU model.

For downstream builds of QEMU, we would like to be able to disable
these legacy CPUs in the build. Thus add a CONFIG switch that can be
used to disable them (and old machine types that use them by default).

Message-Id: <20240614125019.588928-1-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>
(cherry picked from commit d6a7c3f44cf3f60c066dbf087ef79d4b12acc642)

Patch-name: kvm-target-s390x-Add-a-CONFIG-switch-to-disable-legacy-C.patch
Patch-id: 28
Patch-present-in-specfile: True