tests/migration-tests: Cover postcopy failure on reconnect
RH-Author: Juraj Marcin <None>
RH-MergeRequest: 408: migration: New postcopy state, and some cleanups [rhel-9.4.z]
RH-Jira: RHEL-55457
RH-Acked-by: Peter Xu <peterx@redhat.com>
RH-Acked-by: Prasad Pandit <None>
RH-Commit: [12/12] a4b9379ab60243f2285855c9330fa4a36a5e29d4

Make sure there will be an event for postcopy recovery, irrelevant of
whether the reconnect will success, or when the failure happens.

The added new case is to fail early in postcopy recovery, in which case it
didn't even reach RECOVER stage on src (and in real life it'll be the same
to dest, but the test case is just slightly more involved due to the dual
socketpair setup).

To do that, rename the postcopy_recovery_test_fail to reflect either stage
to fail, instead of a boolean.

Reviewed-by: Fabiano Rosas <farosas@suse.de>
Signed-off-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Fabiano Rosas <farosas@suse.de>

(cherry picked from commit 6cf56a87baf8b99c4296a943d220eb8276ca035a)

Conflicts:
	tests/qtest/migration-test.c: migration_test_add() helper
		function is not yet present, use qtest_add_func()
		instead

Signed-off-by: Juraj Marcin <jmarcin@redhat.com>